almost done with the is_complete function
I sorted out the returns, it seems rust prefers that you don't deal with data in a loop and then give it back to the function scope, so it must return from inside the loop. This had to do with lifetimes but the solution there sought to use them in a way unintended for safetys sake, I guess.

After testing, I found out that the original partial/complete algorithm ate shit when it encountered the Unconfirmed*.crdownload files, so I have figured out a way, I think, to continue looping and only return false when the items in the directory have been exhausted. So that is what I will try to pass the tests, and then I will either work out the online test for the function or move on to the next function. I also have not done a flow model of this yet, need to figure out how to make a call graph, and these are oddly help I actually need much earlier in Python because it doesnt offer as much feedback. That said they are probably next as well in order to help me plan the main function.